+++ openrouteservice/src/main/java/org/heigit/ors/matrix/algorithms/AbstractMatrixAlgorithm.java
@@ -18,12 +18,14 @@ import com.graphhopper.routing.util.FlagEncoder;
 import com.graphhopper.routing.weighting.Weighting;
 import com.graphhopper.storage.Graph;
 import org.heigit.ors.matrix.MatrixRequest;
+import org.heigit.ors.services.matrix.MatrixServiceSettings;
 
 public abstract class AbstractMatrixAlgorithm implements MatrixAlgorithm {
   protected GraphHopper graphHopper;
   protected Graph graph;
   protected FlagEncoder encoder;
   protected Weighting weighting;
+  protected int maxVisitedNodes = Integer.MAX_VALUE;
   
   public void init(MatrixRequest req, GraphHopper gh, Graph graph, FlagEncoder encoder, Weighting weighting)
   {
@@ -31,5 +33,6 @@ public abstract class AbstractMatrixAlgorithm implements MatrixAlgorithm {
 	  this.graph = graph;
 	  this.encoder = encoder;
 	  this.weighting = weighting;
+	  this.maxVisitedNodes = MatrixServiceSettings.getMaximumVisitedNodes();
   }
 }